+++ test/validationRules/required-fields.js
@@ -7,8 +7,11 @@ const requiredFieldsTestCases = {
   pass: [
     "const x = gql`query { allFilms { films { title } } }`",
     "const x = gql`query { stories { id comments { text } } }`",
-    "const x = gql`query { greetings { id, hello, foo } }`",
-    "const x = gql`query { greetings { hello ... on Greetings { id } } }`"
+    "const x = gql`query { greetings { id, hello, hi } }`",
+    "const x = gql`query { greetings { id ... on Greetings { hello } } }`",
+    "const x = gql`fragment Name on Greetings { id hello }`",
+    "const x = gql`fragment Id on Node { id ... on NodeA { fieldA } }`",
+    "const x = gql`query { nodes { id ... on NodeA { fieldA } } }`",
   ],
   fail: [
     {
@@ -31,14 +34,52 @@ const requiredFieldsTestCases = {
     },
     {
       code:
-        "const x = gql`query { greetings { hello ... on Greetings { foo } } }`",
+        "const x = gql`query { greetings { hello ... on Greetings { hello } } }`",
       errors: [
         {
           message: `'id' field required on 'greetings'`,
           type: "TaggedTemplateExpression"
         }
       ]
-    }
+    },
+    {
+      code: 'const x = gql`query { greetings { hello ...GreetingsFragment} }`',
+      errors: [
+        {
+          message: `'id' field required on 'greetings'`,
+          type: 'TaggedTemplateExpression',
+        },
+      ],
+    },
+    {
+      code: 'const x = gql`fragment Name on Greetings { hello }`',
+      errors: [
+        {
+          message: `'id' field required on 'fragment Name on Greetings'`,
+          type: 'TaggedTemplateExpression',
+        },
+      ],
+    },
+    {
+      code:
+        "const x = gql`query { greetingOrStory { ... on Greetings { id } ... on Story { comments { text } } } }`",
+      errors: [
+        {
+          message: `'id' field required on '... on Story'`,
+          type: "TaggedTemplateExpression"
+        }
+      ]
+    },
+    {
+      code:
+        "const x = gql`query { nodes { ... on NodeA { id fieldA } ... on NodeB { id fieldB }}}`",
+      errors: [
+        {
+          message: `'id' field required on 'nodes'`,
+          type: "TaggedTemplateExpression"
+        }
+      ]
+    },
   ]
 };
 
@@ -52,7 +93,7 @@ let options = [
   }
 ];
 
-ruleTester.run("testing required-fields rule", rules["required-fields"], {
+ruleTester.run("testing required-fields rule with env", rules["required-fields"], {
   valid: requiredFieldsTestCases.pass.map(code => ({
     options,
     parserOptions,
@@ -74,7 +115,7 @@ options = [
     requiredFields: ["id"]
   }
 ];
-ruleTester.run("testing required-fields rule", rules["required-fields"], {
+ruleTester.run("testing required-fields rule without env", rules["required-fields"], {
   valid: requiredFieldsTestCases.pass.map(code => ({
     options,
     parserOptions,